fix Select Components for the react-hook-form
