+++ components/Form/SearchForm.tsx
@@ -1,10 +1,18 @@
 import { FormEvent, memo } from 'react';
-import Select from '../Select';
+import { Select } from '../Select';
 import { SearchInput } from '../Input';
 import { RiSearchLine } from 'react-icons/ri';
 
 import styles from './SearchForm.module.scss';
 import { useFindMatePostSearchContext } from '@/context/FindMatePost';
+
+const Divide = () => <hr className={styles.divide} />;
+
+const SearchButton = ({ onClick }: { onClick: () => void }) => (
+  <div className={styles.icon} onClick={onClick}>
+    <RiSearchLine />
+  </div>
+);
 
 const SearchForm = () => {
   const { searchCategory, form, onChange, handleSearch } =
@@ -12,7 +20,7 @@ const SearchForm = () => {
 
   const { q, category } = form;
 
-  const onSubmit = (e: FormEvent) => {
+  const onSubmit = (e: FormEvent<HTMLFormElement>) => {
     e.preventDefault();
     handleSearch();
   };
@@ -29,16 +37,14 @@ const SearchForm = () => {
             noOutline
           />
         </div>
-        <hr className={styles.divide} />
+        <Divide />
         <SearchInput
           placeholder="search..."
           name="q"
           value={q}
           onChange={onChange}
         />
-        <div className={styles.icon} onClick={handleSearch}>
-          <RiSearchLine />
-        </div>
+        <SearchButton onClick={handleSearch} />
       </div>
     </form>
   );
