+++ components/Select/Select.module.scss
@@ -1,6 +1,7 @@
 .select {
   width: 100%;
   border: 1px solid rgba(0, 0, 0, 0.3);
+  border-radius: 3px;
   background-color: transparent;
   font-size: 1rem;
   text-align: center;
